Article section/ Admin mode -> when deleting an article section the image files from public folder are also deleted, when updating an article section the image files are also updated in the public folder

+++ server/controllers/articleSectionController.js
@@ -45,10 +45,42 @@ const createArticleSection = async(req,res) => {
 
 const updateArticleSection = async(req,res) => {
     const {id} = req.params
+    const { title } = req.body;
+    
+
     try {
+        const currentArticleSection = await ArticleSection.findById(id)
+
+        const { image, imageClicked } = req.files;
+        const imageFilename = image[0].filename
+        const imageClickedFilename = imageClicked[0].filename
+
+        if (!currentArticleSection) {
+            return res.status(404).json({ error: "Article section not found" });
+        }
+
+        const currentArticleSectionImageClicked = currentArticleSection.imageClicked
+        const currentArticleSectionImage = currentArticleSection.image
+
+
+        if (image && currentArticleSectionImage) {
+            const imagePath = path.join('public', currentArticleSectionImage);
+            fs.unlinkSync(imagePath);
+            
+        }
+
+        if (imageClicked && currentArticleSectionImageClicked) {
+            const imageClickedPath = path.join('public', currentArticleSectionImageClicked);
+            fs.unlinkSync(imageClickedPath);
+         
+        }
+
         const articleSection = await ArticleSection.findOneAndUpdate({_id: id},{
-            ...req.body
-        })
+            title, image: imageFilename, imageClicked: imageClickedFilename
+        },
+        { new: true }
+        )
+
         res.status(200).json(articleSection)
     } catch (err) {
         res.status(400).json({error: err.message})
@@ -59,9 +91,33 @@ const updateArticleSection = async(req,res) => {
 
 const deleteArticleSection = async(req,res) => {
     const {id} = req.params
+   
     try {
+        const currentArticleSection = await ArticleSection.findById(id)
+        if (!currentArticleSection) {
+            return res.status(404).json({ error: "Article section not found" });
+        }
+
+        const currentArticleSectionImageClicked = currentArticleSection.imageClicked
+        const currentArticleSectionImage = currentArticleSection.image
         const articleSection = await ArticleSection.findOneAndDelete({_id: id})
+
+
+        if (currentArticleSectionImage) {
+            const imagePath = path.join('public', currentArticleSectionImage);
+                fs.unlinkSync(imagePath);
+            
+        }
+
+        if (currentArticleSectionImageClicked) {
+            const imageClickedPath = path.join('public', currentArticleSectionImageClicked);
+                fs.unlinkSync(imageClickedPath);
+         
+        }
+
+
         res.status(200).json(articleSection)
+
     } catch (err) {
         res.status(400).json({error: err.message})
     }
